Remove requirement that context must be the first line
This seems to be more a restriction of implementations rather than a requirement for this spec.
JSON does not guarantee the order of fields. DID-spec could only require that context is the first line of a serialization.

Axel
This was discussed month ago and merged but now this issue is back. Why?