Swap relying party for requesting party Will fix #248
In line with the WG's [resolution on 2020-06-30](https://www.w3.org/2019/did-wg/Meetings/Minutes/2020-06-30-did#resolution1), this minimal PR swaps the 3 instances of 'replying party' for 'requesting party'. If merged, Issue #248 can be closed. The closely related issue in the Use Cases document, ([issue 39](https://github.com/w3c/did-use-cases/issues/39) in that repo) has been closed and the UCR now consistently refers to requesting parties (actually, it refers to Requesting Parties, with upper case R and P, which is a difference from the core spec but probably not something to worry about too much).